Add files via upload
This game contains 26 brief cases that each of them contains random mount of money up to 1 million dollars. at the beginning of the game, the contestant chooses one briefcase to be his, without knowing what amount it contains. 
He then selects more and more briefcases whose amount is revealed. As briefcases with lower amounts are revealed, the bank offers him from time to time bigger amount , and opposite that he can take and withdraw from the game. The goal is to leave with as much money as possible up to 1 million $.